+++ insty-frontend/src/app/creator/videos/_component/manage/VideoEdit.tsx
@@ -1,5 +1,6 @@
-import Image from "next/image";
-import React, { useState } from "react";
+import React from "react";
+import VideoForm from "../common/VideoForm";
+import { UploadformData } from "@/app/types";
 
 function VideoEdit({
 	videoId,
@@ -8,249 +9,39 @@ function VideoEdit({
 	videoId: number;
 	onBack: () => void;
 }) {
-	const [tags, setTags] = useState<string[]>([]);
-	const [tagInput, setTagInput] = useState("");
+	const dummyInitialData = {
+		link: "",
+		thumbnail: "",
+		title: "파이썬 개발 환경 설치 가이드",
+		description:
+			"이 영상은 파이썬 개발 환경을 Windows 10/11에서 설치하는 방법을 안내합니다.",
+		price: 19900,
+		tags: ["파이썬", "개발환경", "설치", "윈도우"],
+		environments: [
+			{ value: "Windows 10 / 11 환경", support: "지원" },
+			{ value: "MacOS 환경", support: "미지원" },
+		],
+		coreContents: [
+			"파이썬 공식 사이트 접속",
+			"설치 파일 다운로드 및 실행",
+			"환경 변수 설정",
+			"설치 확인 및 테스트",
+		],
+	};
 	console.log(videoId);
 
-	const [environments, setEnvironments] = useState([
-		{ value: "Windows 10 / 11 환경", support: "지원" },
-	]);
-
-	const [coreContents, setCoreContents] = useState<string[]>([
-		"파이썬 개발 환경 설치 (Windows 기준)",
-	]);
-
-	const handleAddTag = () => {
-		const val = tagInput.trim();
-		if (val && !tags.includes(val)) {
-			setTags([...tags, val]);
-			setTagInput("");
-		}
-	};
-
-	const handleRemoveTag = (idx: number) => {
-		setTags(tags.filter((_, i) => i !== idx));
-	};
-
-	const handleEnvChange = (
-		idx: number,
-		key: "value" | "support",
-		value: string
-	) => {
-		const arr = [...environments];
-		arr[idx][key] = value;
-		setEnvironments(arr);
-	};
-	const handleAddEnv = () =>
-		setEnvironments([...environments, { value: "", support: "지원" }]);
-	const handleRemoveEnv = (idx: number) =>
-		setEnvironments(environments.filter((_, i) => i !== idx));
-
-	const handleCoreChange = (idx: number, value: string) => {
-		const arr = [...coreContents];
-		arr[idx] = value;
-		setCoreContents(arr);
-	};
-
-	const handleAddCore = () => setCoreContents([...coreContents, ""]);
-	const handleRemoveCore = (idx: number) => {
-		setCoreContents(coreContents.filter((_, i) => i !== idx));
+	const handleSubmit = (formData: UploadformData) => {
+		console.log(formData);
 	};
 
 	return (
-		<div>
-			<h2 className="font-bold text-3xl mb-4">콘텐츠 수정</h2>
-			<div className="flex w-full gap-9">
-				<div className="flex flex-col w-2/5 min-w-[220px] max-w-[350px]">
-					<label className="block text-xl font-semibold mb-1">영상 링크</label>
-					<div className="mb-4">
-						<textarea
-							className="w-full h-36 bg-gray-scale-100 rounded-2xl p-2 text-lg resize-none flex flex-shrink-0"
-							placeholder="업로드 하실 영상의 링크를 붙여넣기 해주세요."
-							rows={3}
-						/>
-					</div>
-					<div className="mb-2 w-full h-36 bg-gray-scale-100 rounded-2xl flex items-center justify-center relative">
-						<button className="absolute top-1 right-2 text-black-500">✕</button>
-					</div>
-					<div className="flex flex-col gap-3 mb-2 justify-center align-middle text-center">
-						<button className="flex-1 py-2 bg-primary-blue-600 text-white rounded cursor-pointer">
-							썸네일 선택
-						</button>
-						<button className="flex-1 py-2 border !border-primary-blue-600 text-primary-blue-600 rounded cursor-pointer">
-							실습 자료 파일 선택
-						</button>
-						<button className="text-secondary-red-300 text-lg cursor-pointer">
-							<span>업로드 영상 삭제</span>
-							<span className="text-lg">🗑️</span>
-						</button>
-					</div>
-				</div>
-
-				<div className="flex-1 flex flex-col gap-4">
-					<div>
-						<label className="block text-xl font-semibold mb-1">제목</label>
-						<div className="flex gap-2">
-							<input
-								className="flex-1 bg-gray-scale-100 rounded-2xl p-2 text-black-100"
-								placeholder="설치 가이드 주제 입력"
-							/>
-							<button className="px-3 py-1 bg-gray-100 border border-gray-200 rounded-2xl text-md cursor-pointer">
-								AI추천 문장입력
-							</button>
-						</div>
-					</div>
-
-					<div className="flex gap-4 mt-4">
-						<div className="flex-1">
-							<label className="block text-xl font-semibold mb-1">대상자</label>
-							<input
-								className="w-full bg-gray-scale-100 rounded-2xl p-2 text-black-100"
-								placeholder="예: 파이썬 개발 환경 설치가 처음인 초보자"
-							/>
-						</div>
-						<div className="flex-1">
-							<label className="block text-xl font-semibold mb-1">가격</label>
-							<input
-								className="w-full bg-gray-scale-100 rounded-2xl p-2 text-black-100"
-								placeholder="예: 199,990"
-							/>
-						</div>
-					</div>
-
-					<div className="mt-4">
-						<label className="block text-xl font-semibold mb-1">설명</label>
-						<div className="flex gap-2">
-							<textarea
-								className="flex-1 bg-gray-scale-100 rounded-2xl p-2 text-black-100 resize-none"
-								rows={6}
-								placeholder="설명 내용 입력"
-							/>
-							<button className="px-3 py-1 bg-gray-100 border border-gray-200 rounded-2xl text-md cursor-pointer">
-								AI추천 요약입력
-							</button>
-						</div>
-					</div>
-
-					<div className="mt-4">
-						<label className="block text-xl font-semibold mb-1">
-							설치 환경 체크리스트
-						</label>
-						<div className="flex flex-col gap-2">
-							{environments.map((env, idx) => (
-								<div key={idx} className="flex gap-2 items-center">
-									<input
-										className="flex-1 bg-gray-scale-100 rounded-2xl p-2 text-black-100"
-										value={env.value}
-										onChange={(e) =>
-											handleEnvChange(idx, "value", e.target.value)
-										}
-										placeholder="환경 입력"
-									/>
-									<select
-										className="border border-gray-scale-300 rounded p-2 text-lg"
-										value={env.support}
-										onChange={(e) =>
-											handleEnvChange(idx, "support", e.target.value)
-										}
-									>
-										<option>지원</option>
-										<option>미지원</option>
-									</select>
-									<button
-										className="text-gray-400 ml-2"
-										onClick={() => handleRemoveEnv(idx)}
-									>
-										✕
-									</button>
-								</div>
-							))}
-							<button
-								className="self-center px-5 py-2 border border-gray-scale-300 rounded text-lg mt-1 hover:bg-gray-scale-300"
-								onClick={handleAddEnv}
-							>
-								더 입력하기 +
-							</button>
-						</div>
-					</div>
-
-					<div className="mt-4">
-						<label className="block text-xl font-semibold mb-1">
-							핵심 전달이 되는 핵심 내용
-						</label>
-						<div className="flex flex-col gap-2">
-							{coreContents.map((content, idx) => (
-								<div key={idx} className="flex gap-2 items-center">
-									<input
-										className="flex-1 bg-gray-scale-100 rounded-2xl p-2 text-black-100"
-										value={content}
-										onChange={(e) => handleCoreChange(idx, e.target.value)}
-										placeholder="핵심 내용 입력"
-									/>
-									<button	
-										className="text-gray-400 ml-2"
-										onClick={() => handleRemoveCore(idx)}
-									>
-										✕
-									</button>
-								</div>
-							))}
-							<button
-								className="self-center px-5 py-2 border border-gray-scale-300 rounded text-lg mt-1 hover:bg-gray-scale-300"
-								onClick={handleAddCore}
-							>
-								더 입력하기 +
-							</button>
-						</div>
-					</div>
-
-					<div className="mt-4">
-						<label className="block text-xl font-semibold mb-1">태그</label>
-						<div className="flex gap-2">
-							<input
-								className="flex-1 bg-gray-scale-100 rounded-2xl p-2 text-black-100"
-								value={tagInput}
-								onChange={(e) => setTagInput(e.target.value)}
-								onKeyDown={(e) =>
-									e.key === "Enter" && (e.preventDefault(), handleAddTag())
-								}
-								placeholder="태그 입력 후 Enter"
-							/>
-							<button
-								className="px-3 py-1 border border-gray-scale-300 rounded text-lg cursor-pointer"
-								onClick={handleAddTag}
-							>
-								추가
-							</button>
-						</div>
-						<div className="flex flex-wrap gap-2 mt-2">
-							{tags.map((tag, idx) => (
-								<span
-									key={tag}
-									className="px-4 py-2 rounded-full flex items-center text-lg border border-primary-blue-600"
-								>
-									{tag}
-									<button
-										onClick={() => handleRemoveTag(idx)}
-									>
-										<Image src="/cancel.svg" alt="취소" width={16} height={16}  className="ml-2"/>
-									</button>
-								</span>
-							))}
-						</div>
-					</div>
-
-					<div className="flex justify-end mt-4">
-						<button
-							className="px-8 py-2 bg-primary-blue-600 text-white rounded text-lg"
-							onClick={onBack}
-						>
-							수정 완료
-						</button>
-					</div>
-				</div>
-			</div>
-		</div>
+		<VideoForm
+			subject="콘텐츠 수정"
+			initialData={dummyInitialData}
+			onSubmit={handleSubmit}
+			submitText="수정하기"
+			onBack={onBack}
+		/>
 	);
 }
