+++ insty-frontend/src/app/creator/videos/_component/upload/VideoUpload.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import VideoForm from '../common/VideoForm'
+
+function VideoUpload() {
+  return (
+    <VideoForm subject="영상 업로드" onSubmit={() => {}} onBack={() => {}} />
+  )
+}
+
+export default VideoUpload
